Deque debería ser una cola de futures. O usar las virtudes de java 8 ;/

+++ homework-models/src/main/java/businessobjects/Account.java
@@ -1,10 +1,12 @@
 package businessobjects;
 
+import java.math.BigDecimal;
+
 import models.AbstractAccount;
 
 public class Account extends AbstractAccount {
 	
-	public boolean isDebitLessThanActualCredit(Long montoDebito) {
+	public boolean isDebitLessThanActualCredit(BigDecimal montoDebito) {
 	
 		return this.getCredit().compareTo(montoDebito) >= 0;
 	};
